Implemented ScapSourceBinaries using Indexer to save database

+++ Patch Data Store/FileStore.py
@@ -5,14 +5,18 @@ import dircache
 import shutil
 import string
 
+import Indexer
+
 class FileStore:
 	DebugLevel = 0
-	def __init__(self, OutputDir, target_binaries_folder ):
-		self.OutputDir = OutputDir
+	def __init__(self, source_binaries_folder, target_binaries_folder, indexer_database_name = 'index.db' ):
+		self.SourceBinariesFolder = source_binaries_folder
 		self.TargetBinariesFolder = target_binaries_folder
+		self.IndexerDatabaseName = indexer_database_name
+		self.IndexerDatabase = Indexer.Database( self.IndexerDatabaseName )
 
 	def ExtractFile( self, Filename ):
-		popen2.popen2( Filename + " /x:"+self.OutputDir + " /quiet" )
+		popen2.popen2( Filename + " /x:"+self.SourceBinariesFolder + " /quiet" )
 		
 	def QueryFile( self, filename ):
 		VersionInfo = {}
@@ -34,33 +38,53 @@ class FileStore:
 						print "\t", s, value
 		return VersionInfo
 
-	def ScrapOutput( self, dirname = None ):
+	def ScapSourceBinaries( self, dirname = None ):
 		if not dirname:
-			dirname = self.OutputDir
+			dirname = self.SourceBinariesFolder
 
 		for file in dircache.listdir( dirname ):
 			full_path = os.path.join( dirname, file )
 			if os.path.isdir( full_path ):
-				self.ScrapOutput( full_path )
+				self.ScapSourceBinaries( full_path )
 			else:
-				print full_path
+				if self.DebugLevel > 2:
+					print full_path
 				version_info = self.QueryFile( full_path )
+				filename = os.path.basename( full_path )
 				
 				if len(version_info) > 0:
-					print version_info
-					filename = os.path.basename( full_path )
-					target_directory = os.path.join( self.TargetBinariesFolder, version_info['CompanyName'], filename , string.replace( version_info['FileVersion'], ':', '_' ) )
-					target_full_path = os.path.join( target_directory, filename )
+					if self.DebugLevel > 2:
+						print version_info
+					
+					if self.SourceBinariesFolder != self.TargetBinariesFolder:
+						target_directory = os.path.join( self.TargetBinariesFolder, version_info['CompanyName'], filename , string.replace( version_info['FileVersion'], ':', '_' ) )
+						target_full_path = os.path.join( target_directory, filename )
 
-					if not os.path.isdir( target_directory ):
-						os.makedirs( target_directory )
-					print 'Copy to ',target_full_path
-					shutil.copyfile( full_path, target_full_path )
-		
+						if not os.path.isdir( target_directory ):
+							os.makedirs( target_directory )
+						print 'Copy to ',target_full_path
+						shutil.copyfile( full_path, target_full_path )
+					#TODO: Put to the Index Database
+					print full_path, version_info
+					operating_system = 'Windows XP'
+					patch_identifier = ''
+					service_pack = ''
+
+					self.IndexerDatabase.Add( 
+						operating_system, 
+						service_pack, 
+						filename, 
+						version_info['CompanyName'], 
+						version_info['FileVersion'], 
+						patch_identifier, 
+						full_path
+					)
+
 	def RemoveOutput( self, dirname = None ):
-		if not dirname:
-			dirname = self.OutputDir
-		shutil.rmtree( dirname )
+		if self.SourceBinariesFolder != self.TargetBinariesFolder:
+			if not dirname:
+				dirname = self.SourceBinariesFolder
+			shutil.rmtree( dirname )
 
 		"""
 		for file in dircache.listdir( dirname ):
@@ -121,11 +145,13 @@ class Win32VerTest(unittest.TestCase):
 
 if __name__=='__main__':
 	#unittest.main()
-	OutputDir = "Out"
+	SourceBinariesFolder = "Out"
+	
+	SourceBinariesFolder = r"T:\mat\Projects\Binaries\Windows XP"
 	TargetBinariesFolder = r"T:\mat\Projects\Binaries\Windows XP"
 	PatchBinary = r'Patches\WindowsXP-KB950762-x86-ENU.exe'
 
-	file_store = FileStore( OutputDir, TargetBinariesFolder )
-	file_store.ExtractFile( PatchBinary )
-	file_store.ScrapOutput()
-	file_store.RemoveOutput()
+	file_store = FileStore( SourceBinariesFolder, TargetBinariesFolder )
+	#file_store.ExtractFile( PatchBinary )
+	file_store.ScapSourceBinaries()
+	#file_store.RemoveOutput()
